Rename vanilla option module and add payoff tests
